split tem component for a part fix css code
