+++ components/Card/Card.module.css
@@ -8,18 +8,14 @@
   object-fit: cover;
   width: 100%;
   height: var(--card-img-height);
-}
-@media screen and (max-width: 400px) {
-  .image {
+  @media screen and (max-width: 400px) {
     height: var(--card-img-height-mobile);
   }
 }
 
 .body {
   padding: 24px;
-}
-@media screen and (max-width: 400px) {
-  .body {
+  @media screen and (max-width: 400px) {
     padding: 16px;
   }
 }
